Remove the limit of state machines in Bluetooth BLE profiles
The limit of state machines was added to avoid DoS attack. However, it is not necessary because the state machines are created only when a device is connected. The number of connected BLE devices is limited by the Bluetooth controller and system property. Therefore, the limit of state machines is not necessary.
This CL removes the limit of state machines in Bluetooth BLE profiles.

CRs-Fixed: 4059944
Bug: 316982710
Change-Id: I93322aa5982a2f5a72fb9fa37926ee4164461cb8
Flag: Exempt, max number of BLE devices are guarded by the controller
Test: m packages/modules/Bluetooth
(cherry picked from commit 26ecf9b8d5d0bf1f790b592fb2f2488b2b17a4d6)