GATT: GATT Server Disconnect is failing.
Issue: GATT server disconnection callback is not being sent to the
application. When BTA_GATTS_CLOSE_EVT event is triggered, it does not
perform any callback to notify the application about the GATT server
disconnection.

Fix: Implemented changes to ensure the application receives a callback
on the BTA_GATTS_CLOSE_EVT event.

CRs-Fixed: 4034704
Change-Id: I004892963605e2005312d5fe210576c017a1cf64