GATT: Allow server disconnection only in BT_ON and BLE_ON state.
Issue: The server disconnection and Bluetooth off events
are occurring simultaneously. During BT off the Native Interface
is cleared. As a result, when the server disconnection thread
attempts to access the Native Interface, it encounters a null
exception due to the interface being cleared

Fix: To address this issue, we are preventing server disconnection
when the Bluetooth state is OFF.

CRs-Fixed: 4038750
Change-Id: Ieefbc407be9805bcd35b26f043ba0b1dbb59079b