HFP : Change inband ringtone based with active device.
Description: Previously when there is no LE Audio, if there is 1
connected HFP device, it will always be active. But with LE Audio,
it might be connected but not active. In such case, phone should
notify the remote that in band ringtone will not be played.
It should use out band ringtone.

Change-Id: Iabe14acee436503665d9128e87cba14909c5bae5
CRs-Fixed: 4032858