LEA : set audio_sender_state to READY_TO RELEASE
Issue:
An incorrect state value is being set for audio_sender_state when disconnecting the device during active streaming.

When the disconnection is triggered by the Remote Device (RD), the Device Under Test (DUT) proceeds to release the stream without first setting audio_sender_state to READY_TO_RELEASE. This causes the group to become inactive before the disconnection is properly handled.

Fix:
Since the DUT is responsible for stopping or releasing the stream, it should do so gracefully. To ensure proper disconnection, update the audio_sender_state to READY_TO_RELEASE during the disconnection process if streaming is active.

Change-Id: If298d20fbfba5d49e9f7157ad0cd44b02b2a416a
CRs-Fixed: 4161353