Correct mHasFallback when in-active device gets disconnected
Use-case:
1. Pair & connect RD1 and play music.
2. Connect RD2 and make it active.
3. Disconnect RD1 (inactive device) and reconnect.
4. Disconnect RD2 (inactive device) and reconnect.
5. Repeat steps 3 & 4 at random.

Result: When inactive device is disconnected & reconnected,
it becomes active but audio is in paused state in DUT.

Root cause: mHasFallback is set to false when inactive device disconnected. Expectedly, this flag is only required to set
false when active device disconnects and it will pause music.

Fixed: Set mHasFallback to false when device is disconnected
in stack. And this flag should be reset to true once it updates
device.

Change-Id: Idf61f71c3e9188897da11a0a425934b727baf4c8
CRs-Fixed: 4159811