ble: delay ltk request, if device record is null
Occasionally, the DUT does not respond to the LTK request,
likely because the device record is not yet available at the BTM layer.
This issue arises because the LTK request is received almost immediately
within less than 1 millisecond after the connection is completed.
To address this, delay the LTK request by 50 milliseconds to ensure the
device record is available, allowing the DUT to properly respond to a
remotely initiated LTK request

CRs-Fixed: 4165045
Change-Id: I32bd509805050315d115f824977a7461badde5f5