HFP: Active call indicators are not updated to remote.
When Incoming call is already in progress and a new
connection is initiated, During the process of query
Phone state, the call indicators update is not reaching the
stack. These are getting stored in the queue used for processing
DSDA call indicators of 2 incoming calls.

Modified the conditional check to check previous call state
before queuing them.

CRs-Fixed: 4247604
Change-Id: Icb46c4a7c7745e25a1dbe1e3d011426e5f134dcf